Resolved some previous issues.
