+++ templates/app_base.html
@@ -1,0 +1,16 @@
+{% extends "base.html" %}
+
+{% block sub_nav %}
+<nav class="flex bg-gray-200 text-blue-800 border-b border-gray-300 sm:h-10 items-center">
+  <div class="w-full max-w-6xl relative mx-auto px-6">
+    <div class="flex flex-col sm:flex-row sm:items-center">
+      <a href="/" class="block mr-4 py-1">
+        <h1 class="text-base font-light hover:underline">Today</h1>
+      </a>
+      <a href="/" class="block mr-4 py-1">
+        <h1 class="text-base font-light hover:underline">Courses</h1>
+      </a>
+    </div>
+  </div>
+</nav>
+{% endblock %}
