+++ templates/base.html
@@ -9,7 +9,7 @@
 <body class="min-h-screen flex flex-col">
 
   {% include "nav.html" %}
-
+  {% block sub_nav %}{% endblock %}
   <main class="container mx-auto p-8 flex-grow">
     {% block main %}{% endblock %}
   </main>
